+++ packages/orq-rc/RELEASES.md
@@ -8139,3 +8139,13 @@ Based on:
 - [typescript v4.7.0-rc.57] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.57 - packages/orq-rc
+
+## 2026-04-09 04:27:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.58] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.58] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.58 - packages/orq-rc